requirements.txt
Pillow
reportlab
qrcode
